created skeleton for decorator pattern
