+++ src/Box.java
@@ -8,59 +8,16 @@ import java.awt.*;
  * String name, Int x, Int y, Int width, Int height
  */
 public class Box extends UMLComponent{
-    private String name;
-    private int x;
-    private int y;
-    private int width = 120;
-    private int height = 60;
     public Box(String name, int x, int y) {
-        super(null);
-        this.name = name;
-        this.x = x;
-        this.y = y;
+        super.setName(name);
+        super.setX(x);
+        super.setY(y);
     }
 
+    @Override
     public void paintBox(Graphics g){
-        g.fillRect(this.x + width/2, this.y + height/2, this.width, this.height);
+        g.fillRect(super.getX() - super.getWidth()/2, super.getY() - super.getHeight()/2,
+                super.getWidth(), super.getHeight());
     }
 
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getX() {
-        return x;
-    }
-
-    public void setX(int x) {
-        this.x = x;
-    }
-
-    public int getY() {
-        return y;
-    }
-
-    public void setY(int y) {
-        this.y = y;
-    }
-
-    public int getWidth() {
-        return width;
-    }
-
-    public void setWidth(int width) {
-        this.width = width;
-    }
-
-    public int getHeight() {
-        return height;
-    }
-
-    public void setHeight(int height) {
-        this.height = height;
-    }
 }
