+++ src/UMLComponent.java
@@ -1,11 +1,51 @@
-public class UMLComponent {
-    private final Box box;
+import java.awt.*;
 
-    public Box getBox() {
-        return box;
+public abstract class UMLComponent{
+    private String name;
+    private int x;
+    private int y;
+    private int width = 120;
+    private int height = 60;
+
+    public abstract void paintBox(Graphics g);
+
+    public String getName() {
+        return name;
     }
 
-    public UMLComponent(Box box) {
-        this.box = box;
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getX() {
+        return x;
+    }
+
+    public void setX(int x) {
+        this.x = x;
+    }
+
+    public int getY() {
+        return y;
+    }
+
+    public void setY(int y) {
+        this.y = y;
+    }
+
+    public int getWidth() {
+        return width;
+    }
+
+    public void setWidth(int width) {
+        this.width = width;
+    }
+
+    public int getHeight() {
+        return height;
+    }
+
+    public void setHeight(int height) {
+        this.height = height;
     }
 }
